% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Title : Creation of the normal multivariate distribution function
Description of the function
This function compute for a univariable or mutivariable parameter
the PDF (Probability Distribution Function) for a normal density}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a p x n data matrix with n the number of observations and
p the number of dimensions}

\item{mean}{mean vector}

\item{varcovM}{variance-covariance matrix}

\item{Log}{logical flag for returning the log of the probability density
function. Default is \code{TRUE}.}
}
\value{
a list containing the input matrix x and y the multivariate-Normal probability density function
computed at x
}
\description{
The details
}
\examples{
mvnpdf(x=matrix(1.96), Log=FALSE)
dnorm(1.96)

mvnpdf(x=matrix(rep(1.96, 2), nrow=2, ncol=1), Log=FALSE)

}
